requirements.txt
Flask
google-generativeai
PyMuPDF
python-dotenv